+++ opal/core/views.py
@@ -4,6 +4,7 @@ Re-usable view components
 import functools
 import json
 import datetime
+import warnings
 
 from django.utils.dateformat import format
 from django.http import HttpResponse
@@ -12,6 +13,8 @@ from django.utils.decorators import method_decorator
 from django.core.serializers.json import DjangoJSONEncoder
 from rest_framework import mixins, viewsets
 from django.conf import settings
+
+warnings.simplefilter('once', DeprecationWarning)
 
 
 class LoginRequiredMixin(object):
@@ -40,12 +43,25 @@ def _get_request_data(request):
     return json.loads(data)
 
 
-def _build_json_response(data, status_code=200):
+def json_response(data, status_code=200):
+    """
+    Return a HTTPResponse serializing DATA with the correct serializer
+    """
     response = HttpResponse()
     response['Content-Type'] = 'application/json'
     response.content = json.dumps(data, cls=OpalSerializer)
     response.status_code = status_code
     return response
+
+
+# TODO 0.9.0: Remove this
+def _build_json_response(data, status_code=200):
+    warnthem = """
+    opal.core.views._build_json_response has been re-named to
+opal.core.views.json_response and will be removed in Opal 0.9.0
+"""
+    warnings.warn(warnthem, DeprecationWarning, stacklevel=2)
+    return json_response(data, status_code=status_code)
 
 
 def with_no_caching(view):